Squashed commit of the following:

commit 8d4d89ebe3ba75c3bab09c1eb060116b6479d828
Author: Damian Nowakowski <damian.nowakowski@relativity.com>
Date:   Mon May 17 11:46:41 2021 +0200

    corrections

commit d2b8c52ee631f54ab1709446f2ff3a3aa5a76b35
Author: Damian Nowakowski <damian.nowakowski@relativity.com>
Date:   Mon May 17 11:31:53 2021 +0200

    SqlComparer corrections; Config corrections

+++ Source/Relativity.DataExchange.TestFramework/NUnitExtensions/SqlComparerTestCommand.cs
@@ -103,8 +103,11 @@ namespace Relativity.DataExchange.TestFramework.NUnitExtensions
 				SqlComparerExecutor.RunSqlComparer(leftInputFilePath, rightInputFilePath, resultFile);
 				SqlComparerExecutor.StoreSqlComparerDataForCurrentTest(leftInputFilePath, rightInputFilePath, resultFile, this._testWorkspaceToAdd.ComparerConfigFilePath);
 
-				IntegrationTestHelper.DeleteTestWorkspace(IntegrationTestHelper.IntegrationTestParameters, this.leftInputWorkspaceId);
-				IntegrationTestHelper.DeleteTestWorkspace(IntegrationTestHelper.IntegrationTestParameters, this.rightInputWorkspaceId);
+				if (IntegrationTestHelper.IntegrationTestParameters.DeleteWorkspaceAfterTest)
+				{
+					IntegrationTestHelper.DeleteTestWorkspace(IntegrationTestHelper.IntegrationTestParameters, this.leftInputWorkspaceId);
+					IntegrationTestHelper.DeleteTestWorkspace(IntegrationTestHelper.IntegrationTestParameters, this.rightInputWorkspaceId);
+				}
 
 				testIteration = 0;
 			}
